Fix output schema of TransformObjectCatalogTask
With this change, the output schema becomes independent of the
specific input data, and all filters in the config parameter
filterMap will be mapped into the result schema. NaNs are filled
for those without input data. Filters not specified in the config
will be left out.

This also adds a test for TransformObjectCatalogTask and more
log messages.