Add option to have Mag use calibrated fluxes
Previously, Mag required either a calib object
or calib=None, when a default zero-point would be used.
Now, since the pipe-analysis parquet tables are writing
calibrated fluxes, if we want to compute mags from them,
then there needs to be an option for the Mag calculation
to have a ZP of one; this allows this if calib=False.