Add a detect_isPrimary flag to SFM schema
The coadd level catalogs include a detect_isPrimary flag, useful for culling
catalogs of duplicate (patch/tract overlap), non-deblended parent sources
(deblend_nChild>0), and sky objects. This commit adds an equivalent flag for
single frame processing. Currently, the only two conditions for a source to
be considered primary are that it is not a parent (undeblended) source, and
not a sky source.