Allow a DeferredDatasetHandle to be passed to a Functor
For gen3 compatibility, this allows for a DeferredDatasetHandle
to be be passed to a functor's __call__ method, rather than the
gen2 parquet interface ParquetTable object.  All currently tested
Functor functionality is also tested here for the gen3 model.  Backward
compatibility with gen2 is maintained.  N.B.: the gen3 functor performance
will likely be slower than the gen2 performance; remains to be seen if this
matters at all.