Simplify Gen3 skymap creation tasks.
We now always write skymaps to the same dataset type and collection,
so those are no longer useful arguments.  And we write the dataset
as well as adding the dimensions in BaseSkyMap.register now, so there's
less for these to do.

This also fixes the handling of config.doAppend in MakeDiscreteSkyMap,
which can't have worked properly before because you're just not allowed
to update a skymap in place without changing its name.  The user is
now required to supply an old name that must differ from the new name
when appending.

The makeGen3SkyMap.py script here really should be converted to a
butler subcommand, but I don't want any more scope on this ticket.
It doesn't really need to be backed by a Task, either, given how
simple it is, but that might make it a bit easier to get the right
defualt configuration from obs packages, which is fairly important.