Correct functor logic to work with single-level datarefs, and add tests
Previous updates to Functor to allow for passing a DeferredDatasetHandle
did not have appropriate logic to allow for the handle to point to a dataframe
with only a single-level column index.  This corrects that, and adds corresponding
tests, which were also missing before (which is why this bug was not caught).