Write the QuickFrameMeasurementTask and unit tests.
This is an experimental task with a potentially volatile API
which should not yet be relied upon, i.e. its API is subject to
change without RFC as its primary purpose is in observing scripts
written by T&S.

The tasks goal is to, given a postISR image, quickly and robustly
find the centroid of the brightest star in the image.