Update the idFactory for the deblender input catalog
The current gen3 pipeline does not update the catalog with the
last source id in the mergeDet catalog, so if there were any
merged peaks culled then the next source id starts too low,
resulting in a catalog that is not sorted with some duplicate
entries.

This commit fixes that by notifying the idFactory of the largest
source id.