Propagate optional-output cases in AssembleCoadd and MakeWarp.
AssembleCoadd raises NoWorkFound because it really had nothing to do.

MakeWarp just writes nothing and ends successfully because it might
write one kind of warp and not the other.

This should remove the need for the doWriteEmptyWarps configuration
option, but I haven't attempted to remove or deprecate it yet in order
to avoid backwards compatibility headaches.