Retarget to PsfWcsSelectImagesTask for makeWarp
The PsfWcsSelectImagesTask allows for thresholding on the PSF model
quality based on two (model - source) metrics:

maxEllipResidual: maximum median ellipticity residual
maxScaledSizeScatter: maximum scatter in the scaled size residuals

Any calexps not satisfying the thresholds are considered pour quality
and are not included in the coadds.  This is of fundamental importance
for coadd fidelity, thus we set it as the default for the makeWarp task
here.  Note that the optimal thresholds will be camera/dataset-dependent,
so users should override as appropriate.  The current defaults are those
chosen for HSC data processing.