Validate identifiers in ConfigurableActionStruct
If one adds attributes through a configurable action struct
through the update interface it was possible to use a string
that was not a valid python variable name, only to fail later
when it was read or loaded through a config file. This validates
that the strings are appropriate identifiers.