Fix ConfigurableAction serialization tests.
Tests for serialization were doing nothing, because:

- the config values in the test were set up entirely by the class
  constructor, which is called directly before reading the saved
  config;
- the test was using StringIO.read() instead of StringIO.getvalue(),
  which returns an empty string unless you seek back to the beginning
  of the buffer, and hence we were reading an empty string.

But fixing that was nontrivial, because the classes used in those tests
were defined in the test script, and hence weren't importable (and all
config persistence needs to be able to import the config class and
fields).  So those test classes have been moved to a tests module.