Remove no-longer necessary config gen2/gen3 validation check
The config validation should have gone away when the rest of the gen2
code was purged in the last two weeks.
The remaining uses of `ref_dataset_name` can now refer directly to the
`name` parameter of ReferenceObjectLoader, so that no Science Pipelines
code will need to refer to `ref_dataset_name` any more.