Accommodate WCS to None on fail instead of raise
Instead of just raising, we now set the WCS of an exposure to None if
the single frame processing astrometry fails to find a solution and
allow the possibility for the processing to proceed (controlled via the
requireAstrometry and requirePhotoCal config parameters, more below).
As such, some accommodations must be made for downstream tasks that
require the presence of a WCS to be meaningful.  Here, on an astrometry
fail, if the config parameter requireAstrometry is True, we raise (as
in the previous behaviour).  However, if requireAstrometry is False, we
skip the match persistence (since no matches exists in this case) and we
set the photoCalib to None (again, no matching can be done if there is no
WCS to translate from pixel to sky).  All coord_ra and coord_dec columns
are set to numpy.nan (while leaving the pixel x & y values untouched).
Since, in the case of an astrometry fail, the photometric calibration
cannot proceed, in order to continue processing, the config parameter
requirePhotoCal must also be set to False.  Otherwise, the inability
to perform photometric calibration will result in a raise.

In the coaddition stage, exposures with WCS set to None cannot be
included, so they get "de-selected".

Some testing of this behavior has been added to test_wcsSelectImages.py.
Further testing happens in the ci_hsc repo (where two WCS failures have
been explicitly added).