Allow using external cals even if calexp WCS is None
This also reorders the selection of suitable calexps such that the
selection based on patch overlap is based on the final calibrated
WCS (i.e. if an external calibration is being applied, this one will
be used instead of the calibration based on single frame measurement).

This also includes the deprecation of the prepareCalibratedExposures()
function in favor of a similar version, now with a leading underscore
in the name to make it clear that this function should only be used
internally.  The deprecated version hands off to the new underscored
version for a deprecation period, but will be removed after such time.