Re-add docs that were removed on DM-35670
These might not all be relevant any more, and the paragraphs of text and
Debugging sections should go in the Task Topic docs, but I didn't want
these descriptions to be completely lost. We can sort that out on a
subsequent ticket that moves much of this text up to the task topic
docs.