Add calexpType template to outputs of postprocessing tasks.
Previously, MakeCcdVisitTableTask and MakeVisitTableTask allowed
calexpType to be used for inputs (use case: fakes_calexp) but not
outputs. Making the outputs match makes it possible run both fakes and
non-fakes versions of these tasks in the same pipeline, though the only
difference in output should be a few extra columns in the
fakes version.