Stop using calexpType for datasets no longer in fakes pipeline.
These are all tasks that aren't usefully affected by fake insertion
(at most, rerunning them would change summary stats by including
measurements of fakes that were inserted with the same PSFs, WCS, etc.
that we're trying to summarize, which is just circular).

Unfortunately, we can't actually drop the calexpType connection
template, because that's a config field and those need to be deprecated
to avoid breakage when reading old configs.  That'll have to wait for
DM-37465.