Remove good-for-tests, bad-for-production logic in summary stats.
The check removed here was apparently in place to allow
ComputeSummaryStatsTask to be run in some tests that aren't about the
summary stats run it anyway, under artificial conditions.  But that
check could also turn serious configuration errors into silent failures
at production.  It's better to not run this task in those tests.