Default to psfex for psf measurements
psfex is good enough for most cases and much faster than piff. We'll
use piff in FinalizeCharacterizeTask.

This commit was backported as part of DM-35207, which originally (i.e.
on main) instead removed an override in drp_pipe that prevented the
change on this commit from being applied to the DRP pipelines.