Remove now-defaulted overrides
Appropriate config overrides have been set in the subtasks, so these pipetask
overrides are now either unnecessary (default is now the value that was here),
or incorrect (default is now the value that had been overridden in obs_lsst
and obs_subaru).
This commit is not intended to change pipeline behavior: the combined effect
of this branch across all packages is to cleanup our configs so that the task
defaults are what we believe to be most appropriate for most cases.