Get WCS, PhotoCalib, PSF, and ApCorrMap from visitSummary in MakeWarp.
Since only finalVisitSummary has PSF and ApCorrMap, this adds a new
useVisitSummaryPsf config option that can be set to False to use the
calexp for those instead.  In all other cases we skip any detectors
for which the visit summary doesn't provide what we want, because we
expect that to represent our best calibrations and we don't want to
include detectors without those in the coadd.