Make sure bands exist in DCR subfilter registration test.
It shouldn't be possible to register a subfilter that references a band
that doesn't exist; this is a variant of DM-40818.

*If* one does manage to register such a subfilter, querying for it with
a band constraint (as the CLI script does) is ill-defined - should it
return no records since the band does not exist?  Or should it return
the referencing subfilter recoreds anyway?  The previous test assumed
the latter, but changes to daf_butler on this ticket are changing the
behavior to the former.  Registering the band (by registering an
instrument and hence some physical_filters) first here avoids that
ill-defined behavior, and it should make this test continue to work
when DM-40818 is done.