Add PSF aperture flux delta metric
This measures the maximum difference of a circular aperture flux measured on
the PSF model realized at positions on a sampled grid of un-masked (i.e. "good")
pixels on the detector.  Large differences indicate a problematic fit, so this
metric can be used to thresholded against inclusion of a "bad" exposure in
coadds.