Add PSF aperture correction delta metric
This measures the maximum difference of the computed aperture correction
scaled (divided) by the psfSigma at positions on a sampled grid on the
detector for the base_PsfFlux_instFlux measurement.  Large differences
indicate a problematic fit, so this metric can be used to thresholded
against inclusion of a "bad" exposure in coadds.  The psfSigma scaling
helps normalize against per-band differences and instruments with
different pixel scales/PSF-sampling.