Calibrate background in calibrateImage photometry
The output background should be calibrated to match the calibrated
exposure, so that it can be directly added back to the exposure.
This also identified a bug in the tests:
afwImage.ExposureF(self.truth_exposure.maskedImage) does not make a
deep copy of the pixels, so I was overwriting the truth exposure!