Fix visit summary usage logic in MakeDirectWarpTask.
useVisitSummaryPsf only affects whether the PSF and ApCorr are pulled
from the visit summary; the WCS and PhotoCalib are used
unconditionally.  This is consistent with the docs, and it makes sense
because a visit summary always has a WCS and PhotoCalib, even when
those are identical to the initial_pvi/calexp ones, while the initial
visit summary does not have a PSF or ApCorr (for mostly historical
reasons).