Drop confusing and incorrect WarpedPsf attachment in MakeDirectWarp.
Attaching a new single-detector WarpedPsf to the in-progress warp every
time we add a new exposure makes no sense, and it's actually not the
right WarpedPsf anyway, since it's built from the preliminary WCS and
PSF, not the final ones.  We're already overriding that PSF with a
(correct) CoaddPsf covering all of the warped detectors later.