Apply calibration to warps after applying skyCorr backgrounds.
Background modeling is done on the image without the calibration
applied, and that's how it should be subtracted, too.

This was a clear bug in the old code, but probably a very minor one,
because the photometric calibration is close to constant, and we were
previously only incorrectly multiplying the skyCorr background by the
ratio of the photometric calibration to its average.