Don't expect visit and detector columns in pre-transform source table.
We're already setting these in transfromSourceTable from the data ID,
overriding whatever was there in the pre-transform table, and this lets
transformSourceTable run on calibrateImage's already-calibrated Astropy
output, which lacks them.

Note that I'm making minor adjustments to Source.yaml instead of
switching to the existing configuration for initial_stars because I
don't want to reconfigure downstream analysis tasks (which require
many measurement columns that calibrateImage doesn't run by default)
on this ticket.  I expect the final configuration to be somewhere in
between.