Switch Transform*Table outputs to ArrowAstropy.
This does not include TransformForcedSourceTable, as its string
columns make it a little trickier.

The inputs and calculations are still Pandas; we just convert to
ArrowAstropy at the end.  This gets the default dataset type
definitions in shape with minimal effort.

There is a temporary exception or TransformObjectTable when
multiLevelOutput=True - that still uses DataFrame - but this option is
now depecated (it wasn't being used in production already).