Add check on psf size to MeasurePsfTask
This will catch bad psfs returned from both psfex and piff.

Unfortunately, there are no tests for MeasurePsf, so there's not a good
place to put a test of this new exception (e.g. with a mocked-bad psf).
I checked it with one of the calibrateImage tests, removing the `not`
and seeing that the appropriate exception was raised.