Add thresholds on the model PSF ellipticity in SFP
In order to guard against irregular PSF shapes (e.g the trefoil, trailed,
kettlebell, etc. shapes often encountered during commissioning), impose
thresholds on the measured PSF model on its ellipticity as well as its
unnormalized ellipticity, defined as hypot(Ixx - Iyy, 2Ixy).  These
thresholds can be set per band and a fallback value is used if the
current band is not included in the threshold config dict.  Detectors
where the initial PSF model has values exceeding these thresholds raise
an UnprocessableDataError such that downstream tasks know to skip them
(and thus do not result in processing failures that would waste human
time chasing down).

The values here have been conditioned on the LSSTComCam DP1 dataset and
do a good job at only excludeing truly "bad" PSFs.  The cuts are intentionally
somewhat conservative to allow SFP to continue on PSFs that are more in the
"iffy" catagory.  For science cases that require strict quality of the PSF,
further cuts can be made on the detectors that are included in coaddition.