Add full tract background functionality.
`matchBackgrounds` in its original form matches by warps, i.e.
single detectors.  A more sensible thing to do is to match
backgrounds across the whole focal plane.  This functionality
needed to be added to do this using warped exposures, in tract
coordinates, so this is now added to `backgrounds.py`.  This
commit also includes a partial revision to `matchBackgrounds.py`
using this new functionality, choosing a reference visit ID based
on these background models instead of on individual warps.  Full
functionality has yet to be restored in light of these changes.