Fully calibrate pixels in MakeWarp.
Previously we were applying the spatially-varying part of the
PhotoCalib while leaving the mean (and hence the units) set to some
semi-arbitrary value. This would have been okay (albeit unnecessarily
complex) if we had been setting the PhotoCalib on the warp
accordingly, to a constant with that arbitrary value.  But we were
actually setting it to the PhotoCalib of the input image where that
arbitrary value came from, and getting lucky that downstream code only
evaluated its mean (which was correct) rather than its spatial
variation (which was very very wrong).

The upshot of all this is that warps now have pixels in nJy units and
attach a trivial PhotoCalib that correctly reflects this.  They also
set the FITS BUNIT keyword.