Hips code updates
This encompases allowing the hips code to be run in a parallel
way at its highest order, and running the rest in one quanta,
skipping image creation at the lowest level. This means the
task needs to be run with two different labels in a pipeline with
different configurations. By default it will run the way it always
has.

This also allows hips code to output either webp or png format.