WIP: Enable dynamic thresholding strategy
This implementation is able to much better a wide range of "scenes",
e.g. crowded and/or high nebulasigy fields in addition to the more
"typical" regions, by implementing a dynamic and iterative thresholding
based on the current detection masks tuned to adapt to the current
scene.  In general, higher detection thresholds are requird for non-
standard fields to avoid contiguous detection footprints which leave
no good isolated sources for calibration.