Enable adaptive thresholding detection strategy
This implementation is to allow the single frame processing calibration
to handle a wide range of "scenes", e.g. crowded and/or high nebulosity
ields in addition to the more "typical" regions, by implementing an
adaptive and iterative thresholding based on the current detection masks
and tuning the thresholds to adapt to the current scene.  In general,
higher detection thresholds are requird for non-standard fields to avoid
large contiguous detection footprints which leave too few good isolated
sources for calibration.

To achieve the previous behaviour, the adaptive scheme can be turned off
by setting the do_adaptive_threshold_detection config to False.