Use deconvolved refs to order bands
When a coadd in one band is catastrophically bad then deconvolution
is unable to run, resulting in missing bands in the deconvovled refs.
This caused a failure due to sorting the coadd refs, so now we
sort over the deconvolved refs and check their equivalence with the
subset of matching coadd refs.