Add functionality to classify a science image as trailed
The code gets post-isr science image from butler given a data query.
The image is input into the trained machine learning model after processing.
The model outputs the classification probability and assigns a label for threshold > 0.5.
The code outputs an astorpy table with vists, detector, probability, and label for the image.