Add tasks for ToO pipeline
This includes one task to build templates using archival DECam
data. The data is expected to be on disk and does not use standard
butler methods for input. A text file stores tile boundaries and
filepaths. It also includes a task to mask DETECTED and
DETECTED_NEGATIVE pixels on an image that are connected to sources
from a reference catalog.