Further robustify calibrateImage pedetstal background
This adds an extra safeguard on the pedestal background measurement by
adding an iteration on the measurement, increasing (doubling) the bin size
at each stage.  The bin size starts very small and there is at least one
iteration.  The stopping criterion is if either the change in the cumulative
pedestal between the current and subsequent level is less than 5% of the
current measured (cumulative) pedestal (relative change) or 0.5 "counts"
(absolute change) or if the bin size has reached a value greater than the
size of the exposure's bounding box.