Simplify the creation of "empty" background list
The former version assumed that the "detection" task had a "background"
attribute.  This is not actually guaranteed, and not necessary to achieve
the desired outcome of a backgroundList object which all NaNs.

This also reduces duplication by replacing an independent setting of an
"empty" backgroundList with a call to this internal function.